Combine covid population modules together, to open door to (unimplemented, future) optimization (... if I implement it in a way that allows, can split these back up in to individual modules maybe)
